+++ Backend/src/main/java/com/gameshub/model/request/PhysicalProductRequestDAO.java
@@ -1,10 +1,12 @@
 package com.gameshub.model.request;
 
+import com.gameshub.model.request.image.PhysicalProductRequestImage;
 import com.gameshub.model.user.*;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.*;
+import java.util.Set;
 
 @Data
 @NoArgsConstructor
@@ -12,7 +14,7 @@ import java.time.*;
 @Table(name = "physicalproductrequest")
 public class PhysicalProductRequestDAO extends ProductRequestDAO {
 
-    public PhysicalProductRequestDAO(LocalDate dateReceived, String status, String requestType, String title, int price, String description, LocalDate postDate, int count, String category, SellerDAO seller) {
+    public PhysicalProductRequestDAO(LocalDate dateReceived, String status, String requestType, String title, float price, String description, LocalDate postDate, int count, String category, SellerDAO seller) {
         this.dateReceived = dateReceived;
         this.status = status;
         this.requestType = requestType;
